tooling: refactor commonly used stuff into xeplib.py
